requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
crispy-bootstrap5==0.7
cryptography==38.0.4
defusedxml==0.7.1
Django==3.2.7
django-allauth==0.51.0
django-auto-prefetch==1.4.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.8.1
idna==3.4
oauthlib==3.2.2
Pillow==9.3.0
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.7
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.3
urllib3==1.26.13
